hpo_optuna/requirements.txt
jupyterlab==1.2.3
matplotlib==3.1.2
neptune-client>=0.4.96
numpy==1.17.0
optuna==0.19.0
psutil==5.6.3
tensorflow==2.0.0
